fix: organize api details folder
