+++ src/components/ApiDetails/ApiOpenSpec/KuskExtension.styled.tsx
