+++ src/components/ApiDetails/ApiOpenSpec/OpenApiEditor.styled.tsx
