+++ src/components/ApiDetails/ApiPaths/ApiPaths.tsx
