+++ src/components/ApiDetails/ApiPaths/index.ts
