+++ src/components/ApiDetails/ApiPaths/styled.tsx
